Fixed routing issue
the search results does not work yet but the routing is fixed
