+++ firebase.js
@@ -2,10 +2,11 @@
 import { initializeApp } from "firebase/app";
 import { getAnalytics } from "firebase/analytics";
 import { getAuth } from "firebase/auth";
-import { getFirestore } from "firebase/firestore"; // Import getFirestore
+import { getFirestore } from "firebase/firestore"; 
+
+
 
 const firebaseConfig = {
-  // Your Firebase configuration
   apiKey: "AIzaSyDFyBtrCITLqt5lsYWvdVDTwy8ZPCPPKFU",
   authDomain: "soleswap-d3aac.firebaseapp.com",
   projectId: "soleswap-d3aac",
@@ -19,6 +20,6 @@ const firebaseConfig = {
 const app = initializeApp(firebaseConfig);
 const analytics = getAnalytics(app);
 const auth = getAuth(app);
-const db = getFirestore(app); // Initialize Firestore
+const db = getFirestore(app);
 
-export { app, analytics, auth, db }; // Export db here
+export {app, analytics, auth, db }
