+++ src/App.jsx
@@ -11,31 +11,40 @@ import Brands from '../src/pages/Shoes/Brands.jsx';
 import Kids from '../src/pages/Shoes/Kids.jsx';
 import Mens from '../src/pages/Shoes/Mens.jsx';
 import Womens from '../src/pages/Shoes/Womens.jsx';
-
+import ForgotPassword from '../src/pages/Authentication/ForgotPassword';
+import UserDashboard from '../src/pages/Account/UserDashboard';
+//import SearchResults from '../src/pages/search/SearchResults'; 
 
 function App() {
   return (
     <Router>
       <Routes>
-        <Route exact path="/" element={<Homepage />} />
+        <Route path="/" element={<Homepage />} />
         <Route path="/home" element={<Homepage />} />
-        <Route exact path="/createaccount" element={<CreateAccount />} />
-        <Route exact path="/login" element={<Login />} />
-        <Route exact path="/forgot-password" element={<ForgotPassword />} />
-        <Route exact path="/UserDashboard" element={<UserDashboard />} />
-        <Route exact path="/profile" element={<Profile />} />
-        <Route exact path="/contact" element={<Contact />} />
-        <Route exact path="/about" element={<About />} />
-        <Route exact path="/FAQ" element={<FAQ />} />
-        <Route exact path="/brands" element={<Brands/>} />
-        <Route exact path="/kid's" element={<Kids/>} />
-        <Route exact path="/mens" element={<Mens/>} />
-        <Route exact path="/women's" element={<Womens/>} />
-        <Route path="/displaysearchresults" element={<DisplaySearchResults />} />
+        <Route path="/createaccount" element={<CreateAccount />} />
+        <Route path="/login" element={<Login />} />
+        <Route path="/forgot-password" element={<ForgotPassword />} />
+        <Route path="/userdashboard" element={<UserDashboard />} />
+        <Route path="/profile" element={<Profile />} />
+        <Route path="/contact" element={<Contact />} />
+        <Route path="/about" element={<About />} />
+        <Route path="/faq" element={<FAQ />} />
+        <Route path="/brands" element={<Brands />} />
+        <Route path="/kids" element={<Kids />} />
+        <Route path="/mens" element={<Mens />} />
+        <Route path="/womens" element={<Womens />} />
+
       </Routes>
     </Router>
   );
 }
 
 export default App;
+
+// <Route path="/" element={<SearchResults />} /> *this should go under the womens route path*
+
+
+
+
+
