+++ src/pages/Home/HomePage.jsx
@@ -1,42 +1,28 @@
-import React, { useState } from 'react';
+import React from 'react';
 import Slider from 'react-slick';
-import { useNavigate } from 'react-router-dom'; 
-import DisplaySearchResults from '../DisplaySearchResults/DisplaySearchResults'; 
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import '../Home/home.css';
 
+
+//add photo for search bar 
+
 function Homepage() {
-  const [searchQuery, setSearchQuery] = useState('');
-  const navigate = useNavigate(); // changed this line becuase it was causing errors
+  const carouselItems = [
+    { id: 1, imageUrl: 'src/assets/c3.jpg', text: 'Shoe 1' },
+    { id: 2, imageUrl: 'src/assets/c1.jpg', text: 'Shoe 2' },
+    { id: 3, imageUrl: 'src/assets/c2.jpg', text: 'Shoe 3' },
+  ];
 
+const carouselSettings = {
+dots: true,
+infinite: true,
+speed: 500,
+slidesToShow: 1,
+slidesToScroll: 1,
+};
 
-  //Search results function when a user types something into the search bar and presses the
-  // enter key button it will redirect them tot he displayresultspage
-  
-  const handleSearch = (e) => {
-    if (e.key === 'Enter') {
-      // Redirect to DisplaySearchResults page with search query as URL parameter
-      navigate(`/DisplaySearchResults?query=${searchQuery}`); 
-    }
-  };
-
-
-  const carouselItems = [
-    { id: 1, imageUrl: 'src/assets/c3.jpg', text: 'Shoe 1' },
-    { id: 2, imageUrl: 'src/assets/c1.jpg', text: 'Shoe 2' },
-    { id: 3, imageUrl: 'src/assets/c2.jpg', text: 'Shoe 3' },
-  ];
-
-  const carouselSettings = {
-    dots: true,
-    infinite: true,
-    speed: 500,
-    slidesToShow: 1,
-    slidesToScroll: 1,
-  };
-
-  return (
+return (
     <div id="root">
       {/* Header Section */}
       <div className="header-content">
@@ -44,20 +30,16 @@ function Homepage() {
           <img src="src/assets/SOLE SWAP.png" alt="Sole Swap Logo" className="logo" />
         </div>
 
-        <div className="search-bar">
-          <input
-            type="text"
-            placeholder="Search..."
-            value={searchQuery}
-            onChange={(e) => setSearchQuery(e.target.value)}
-            onKeyPress={handleSearch}
-          />
+        <div className="search-bar"> 
+          <input type="text" placeholder="Search..." />
+         
         </div>
 
         <div className="sign-up-login">
         <button>❤️</button>
           <button> <a href="/createaccount">Sign Up</a></button>
           <button>< a href ="/Login">Login</a></button>
+         
         </div>
       </div>
 
@@ -71,8 +53,8 @@ function Homepage() {
         </ul>
       </nav>
 
-      {/* Carousel Section */}
-      <div className="carousel-container">
+     {/* Carousel Section */}
+     <div className="carousel-container">
         <Slider {...carouselSettings}>
           {carouselItems.map((item) => (
             <div key={item.id}>
@@ -91,20 +73,24 @@ function Homepage() {
       </div>
 
       <div className="start-trade">
-        <button className="start-trading-button">Start Trading</button>
-        <div className="start-trading-background"></div>
-      </div>
+    <button className="start-trading-button">Start Trading</button>
 
-      <footer className="footer">
-        <ul>
-          <li><a href="/about">About Us</a></li>
-          <li><a href="/contact">Contact</a></li>
-          <li><a href="/faq">FAQ</a></li>
-        </ul>
-      </footer>
+    <div className="start-trading-background"></div>
+  </div>
 
-    </div>
-  );
+     
+     
+
+<footer class="footer">
+  <ul>
+    <li><a href="/about">About Us</a></li>
+    <li><a href="/contact">Contact</a></li>
+    <li><a href="/faq">FAQ</a></li>
+  </ul>
+</footer>
+
+</div>
+);
 }
 
 export default Homepage;
