+++ src/pages/search/search.css
@@ -1,0 +1,32 @@
+.shoes-container {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 20px;
+  }
+  
+  .shoe-item {
+    border: 2px solid #000;
+    padding: 10px;
+    margin-bottom: 20px;
+    border-radius: 8px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 200px;
+  }
+  
+  .shoe-item img {
+    margin-bottom: 10px;
+    width: 100px;
+    height: auto;
+  }
+  
+  .search-input {
+    margin-bottom: 10px;
+  }
+  
+  .search-button {
+    margin-left: 10px;
+  }
+
